+++ extensions/wrapper/wrapper-api/src/main/java/de/sovity/edc/ext/wrapper/api/usecase/model/CatalogFilterExpressionLiteralType.java
@@ -16,7 +16,7 @@ package de.sovity.edc.ext.wrapper.api.usecase.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
-@Schema(description = "Value type of an asset selector criterion right expression value", enumAsRef = true)
+@Schema(description = "Value type of a filter expression criterion", enumAsRef = true)
 public enum CatalogFilterExpressionLiteralType {
     VALUE, VALUE_LIST
 }