+++ extensions/wrapper/wrapper-api/src/main/java/de/sovity/edc/ext/wrapper/api/usecase/model/CatalogQuery.java
@@ -19,13 +19,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
 
+import java.util.List;
+
 @Data
 @AllArgsConstructor
 @RequiredArgsConstructor
 @Schema(description = "Catalog query parameters")
 public class CatalogQuery {
     @Schema(description = "Target EDC DSP endpoint URL", requiredMode = Schema.RequiredMode.REQUIRED)
-    private String targetEdc;
+    private String connectorEndpoint;
 
     @Schema(description = "Limit the number of results", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
     private Integer limit;
@@ -33,6 +35,6 @@ public class CatalogQuery {
     @Schema(description = "Offset for returned results, e.g. start at result 2", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
     private Integer offset;
 
-    @Schema(description = "Filter expression for catalog filtering", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
-    private CatalogFilterExpression filterExpression;
+    @Schema(description = "Filter expressions for catalog filtering", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
+    private List<CatalogFilterExpression> filterExpressions;
 }