+++ utils/catalog-parser/src/main/java/de/sovity/edc/utils/catalog/DspCatalogService.java
@@ -30,7 +30,7 @@ public class DspCatalogService {
     private final DspDataOfferBuilder dspDataOfferBuilder;
 
     public DspCatalog fetchDataOffers(String endpoint) throws DspCatalogServiceException {
-        var catalogJson = fetchDcatResponse(endpoint);
+        var catalogJson = fetchDcatResponse(endpoint, QuerySpec.max());
         return dspDataOfferBuilder.buildDataOffers(endpoint, catalogJson);
     }
 
@@ -39,24 +39,10 @@ public class DspCatalogService {
         return dspDataOfferBuilder.buildDataOffers(endpoint, catalogJson);
     }
 
-    private JsonObject fetchDcatResponse(String connectorEndpoint) {
-        var raw = fetchDcatRaw(connectorEndpoint);
-        var string = new String(raw, StandardCharsets.UTF_8);
-        return JsonUtils.parseJsonObj(string);
-    }
-
     private JsonObject fetchDcatResponse(String connectorEndpoint, QuerySpec querySpec) {
         var raw = fetchDcatRaw(connectorEndpoint, querySpec);
         var string = new String(raw, StandardCharsets.UTF_8);
         return JsonUtils.parseJsonObj(string);
-    }
-
-    @SneakyThrows
-    private byte[] fetchDcatRaw(String connectorEndpoint) {
-        return catalogService
-                .requestCatalog(connectorEndpoint, "dataspace-protocol-http", QuerySpec.max())
-                .get()
-                .orElseThrow(DspCatalogServiceException::ofFailure);
     }
 
     @SneakyThrows